# HumioAPI/_build_tmp/documentation/users-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HumioAPI Users",
    "version": "1.0.0"
  },
  "paths": {
    "/api/users/register": {
      "post": {
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserCreatedResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Get current authorized user profile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "summary": "Update current authorized user fields",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get user by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "summary": "Update user fields",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Delete user by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/reset-password": {
      "post": {
        "summary": "Reset user password",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "No Content" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/users/{id}/roles": {
      "get": {
        "summary": "List user roles",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRolesResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "post": {
        "summary": "Add roles to user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRolesRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRolesResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Remove roles from user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRolesRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRolesResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/users/by-email": {
      "get": {
        "summary": "Get user by email",
        "parameters": [
          { "name": "email", "in": "query", "required": true, "schema": { "type": "string", "format": "email" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "List users",
        "parameters": [
          { "name": "skip", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "take", "in": "query", "required": false, "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 100 } },
          { "name": "email", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsersPageResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/forgot-password": {
      "post": {
        "summary": "Send reset code to email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "No Content" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/reset-password": {
      "post": {
        "summary": "Reset password by code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordByTokenRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "No Content" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterUserRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "name": { "type": "string", "nullable": true }
        }
      },
      "PatchUserRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "nullable": true },
          "name": { "type": "string", "nullable": true },
          "phoneNumber": { "type": "string", "nullable": true }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email" }
        }
      },
      "ResetPasswordByTokenRequest": {
        "type": "object",
        "required": ["email", "token", "newPassword"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "token": { "type": "string" },
          "newPassword": { "type": "string", "minLength": 6 }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["newPassword"],
        "properties": {
          "newPassword": { "type": "string", "minLength": 6 }
        }
      },
      "UserRolesRequest": {
        "type": "object",
        "required": ["roles"],
        "properties": {
          "roles": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "UserRolesResponse": {
        "type": "object",
        "required": ["roles"],
        "properties": {
          "roles": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "UserCreatedResponse": {
        "type": "object",
        "required": ["id", "email"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "email": { "type": "string" },
          "name": { "type": "string", "nullable": true }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": ["id", "email", "createdAt"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "email": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastSeen": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "UsersPageResponse": {
        "type": "object",
        "required": ["total", "items"],
        "properties": {
          "total": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      }
    }
  }
}